Added Basic account management features.
- The accounts view can now be closed
- Accounts can be added
- An account can be selected from the account view

Currently the account view creates a new statistics controller as the user clicks an account. This should be refactored to reuse the existing controller or create a new one only on app start.

The code for creating an account is duplicated in the index and the newaccount controller. This needs to be refactored.